requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
kafka-python==2.0.2
pyspark==3.5.0
redis==5.0.1
python-binance==1.0.19
pandas==2.1.3
numpy==1.25.2
python-multipart==0.0.6
pydantic==2.5.0
aiohttp==3.9.1
asyncio-mqtt==0.13.0